move extensionResourceLoader service into platform
