+++ src/vs/platform/extensionResourceLoader/browser/extensionResourceLoaderService.ts
@@ -6,15 +6,17 @@
 import { URI } from 'vs/base/common/uri';
 import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IFileService } from 'vs/platform/files/common/files';
-import { AbstractExtensionResourceLoaderService, IExtensionResourceLoaderService } from 'vs/workbench/services/extensionResourceLoader/common/extensionResourceLoader';
+import { FileAccess, Schemas } from 'vs/base/common/network';
 import { IProductService } from 'vs/platform/product/common/productService';
-import { asTextOrError, IRequestService } from 'vs/platform/request/common/request';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
+import { ILogService } from 'vs/platform/log/common/log';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { CancellationToken } from 'vs/base/common/cancellation';
+import { AbstractExtensionResourceLoaderService, IExtensionResourceLoaderService } from 'vs/platform/extensionResourceLoader/common/extensionResourceLoader';
 
-export class ExtensionResourceLoaderService extends AbstractExtensionResourceLoaderService {
+class ExtensionResourceLoaderService extends AbstractExtensionResourceLoaderService {
+
+	declare readonly _serviceBrand: undefined;
 
 	constructor(
 		@IFileService fileService: IFileService,
@@ -22,21 +24,32 @@ export class ExtensionResourceLoaderService extends AbstractExtensionResourceLoa
 		@IProductService productService: IProductService,
 		@IEnvironmentService environmentService: IEnvironmentService,
 		@IConfigurationService configurationService: IConfigurationService,
-		@IRequestService private readonly _requestService: IRequestService,
+		@ILogService private readonly _logService: ILogService,
 	) {
 		super(fileService, storageService, productService, environmentService, configurationService);
 	}
 
 	async readExtensionResource(uri: URI): Promise<string> {
-		if (this.isExtensionGalleryResource(uri)) {
-			const headers = await this.getExtensionGalleryRequestHeaders();
-			const requestContext = await this._requestService.request({ url: uri.toString(), headers }, CancellationToken.None);
-			return (await asTextOrError(requestContext)) || '';
-		}
-		const result = await this._fileService.readFile(uri);
-		return result.value.toString();
-	}
+		uri = FileAccess.asBrowserUri(uri);
 
+		if (uri.scheme !== Schemas.http && uri.scheme !== Schemas.https) {
+			const result = await this._fileService.readFile(uri);
+			return result.value.toString();
+		}
+
+		const requestInit: RequestInit = {};
+		if (this.isExtensionGalleryResource(uri)) {
+			requestInit.headers = await this.getExtensionGalleryRequestHeaders();
+			requestInit.mode = 'cors'; /* set mode to cors so that above headers are always passed */
+		}
+
+		const response = await fetch(uri.toString(true), requestInit);
+		if (response.status !== 200) {
+			this._logService.info(`Request to '${uri.toString(true)}' failed with status code ${response.status}`);
+			throw new Error(response.statusText);
+		}
+		return response.text();
+	}
 }
 
 registerSingleton(IExtensionResourceLoaderService, ExtensionResourceLoaderService, InstantiationType.Delayed);
