+++ src/vs/platform/extensionResourceLoader/common/extensionResourceLoader.ts
